added a donut chart to show user stats using chart js
