+++ package.json
@@ -28,12 +28,14 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.1",
     "bootstrap": "^5.1.3",
+    "chart.js": "^3.7.0",
     "connect-mongodb-session": "^3.1.1",
     "express": "^4.17.2",
     "express-session": "^1.17.2",
     "mongoose": "^6.1.8",
     "react": "^16.8.0",
     "react-bootstrap": "^2.1.1",
+    "react-chartjs-2": "^4.0.1",
     "react-dom": "^16.8.0",
     "react-elastic-carousel": "^0.11.5",
     "react-router-dom": "^6.2.1",
